+++ tuna/miopen/driver/base.py
@@ -27,6 +27,7 @@
 """Module that encapsulates the DB representation of a Driver cmd"""
 
 from typing import List, Union, Dict, Any
+from re import search
 from abc import abstractmethod
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
@@ -37,8 +38,8 @@ from tuna.dbBase.sql_alchemy import DbSession
 from tuna.utils.db_utility import build_dict_val_key, get_session_val_map
 from tuna.miopen.db.miopen_tables import TensorTable
 from tuna.miopen.db.miopen_tables import ConvolutionConfig
-from tuna.miopen.utils.metadata import TENSOR_PRECISION
-from tuna.miopen.utils.parsing import parse_line
+from tuna.miopen.utils.metadata import TENSOR_PRECISION, DIR_MAP
+from tuna.miopen.utils.parsing import parse_line, get_fds_from_cmd
 from tuna.driver import DriverBase
 
 LOGGER = setup_logger('MIOpenDriver_driver_base')
@@ -53,6 +54,21 @@ class MIOpenDriver(DriverBase):
 
   def __init__(self, line: str = str(), db_obj: ConvolutionConfig = None):
     super().__init__(line, db_obj)
+
+  @abstractmethod
+  def config_set_defaults(self) -> None:
+    """Setting config DB defaults to avoid duplicates through SELECT"""
+    raise NotImplementedError("Not implemented")
+
+  @abstractmethod
+  def set_cmd(self, data_type: str) -> None:
+    """Set cmd based on tensor data type"""
+    raise NotImplementedError("Not implemented")
+
+  @abstractmethod
+  def compose_weight_t(self) -> dict:
+    """Build weight_tensor"""
+    raise NotImplementedError("Not implemented")
 
   def parse_row(self, db_obj: ConvolutionConfig):
     """Abstract/Inference for Overwritting base class function for batch_norm"""
@@ -81,6 +97,20 @@ class MIOpenDriver(DriverBase):
   def get_common_cols() -> List[str]:
     """Returns common MIOpenDriver command line args"""
     return ['wall', 'time', 'iter', 'verify']
+
+  def parse_fdb_key(self, line: str) -> None:
+    """import config attributes from fdb key line"""
+    fds: str
+    direction: str
+    fds, _, direction = get_fds_from_cmd(line)
+    setattr(self, 'direction', DIR_MAP[direction])
+    for key in self.to_dict():
+      if key in fds:
+        setattr(self, key, fds[key])
+
+    pattern_3d = '[0-9]x[0-9]x[0-9]'
+    if search(pattern_3d, line):
+      setattr(self, 'spatial_dim', 3)
 
   def construct_driver(self, line: str) -> bool:
     """Takes a MIOpenDriver cmd or PDB key"""
