+++ tuna/miopen/driver/convolution.py
@@ -27,7 +27,6 @@
 """Module that a convolution MIOpenDriver cmd"""
 
 from typing import Dict, Set, Optional, Any
-from re import search
 from tuna.utils.logger import setup_logger
 from tuna.miopen.driver.base import MIOpenDriver
 from tuna.miopen.utils.metadata import CONV_CONFIG_COLS
@@ -37,7 +36,7 @@ from tuna.miopen.utils.metadata import CONV_2D_DEFAULTS, SUPPORTED_CONV_CMDS, PR
 from tuna.miopen.utils.metadata import CONV_3D_DEFAULTS, TENSOR_COLS
 from tuna.miopen.utils.metadata import TABLE_COLS_CONV_MAP, TENSOR_PRECISION, DIR_MAP
 from tuna.miopen.utils.metadata import DIRECTION, CONV_SKIP_ARGS, INVERS_DIR_MAP
-from tuna.miopen.utils.parsing import get_fd_name, conv_arg_valid, get_fds_from_cmd
+from tuna.miopen.utils.parsing import get_fd_name, conv_arg_valid
 from tuna.miopen.utils.config_type import ConfigType
 
 LOGGER = setup_logger('driver_conv')
@@ -127,20 +126,6 @@ class DriverConvolution(MIOpenDriver):
   def get_layouts(self):
     """Get convolution layouts"""
     return ["in_layout", "out_layout", 'fil_layout']
-
-  def parse_fdb_key(self, line: str) -> None:
-    """import config attributes from fdb key line"""
-    fds: str
-    direction: str
-    fds, _, direction = get_fds_from_cmd(line)
-    setattr(self, 'direction', DIR_MAP[direction])
-    for key in self.to_dict():
-      if key in fds:
-        setattr(self, key, fds[key])
-
-    pattern_3d = '[0-9]x[0-9]x[0-9]'
-    if search(pattern_3d, line):
-      setattr(self, 'spatial_dim', 3)
 
   def parse_driver_line(self, line: str) -> None:
     """Parse MIOpenDriver line"""
